% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.r
\docType{methods}
\name{rcor}
\alias{rcor}
\alias{rcor.gls}
\alias{rcor.asreml}
\alias{print.asremlrmat}
\title{Extract the R-side variance components from a fitted model (asreml or lme)
and format them as matrices.}
\usage{
rcor(obj)

\method{rcor}{gls}(obj)

\method{rcor}{asreml}(obj)

\method{print}{asremlrmat}(x, n = 6, dig = 3, ...)
}
\arguments{
\item{obj}{A fitted model object.}

\item{x}{Object to print}

\item{n}{Size of matrix corner to print}

\item{dig}{Number of digits for printing}

\item{...}{Other arguments (not used)}
}
\value{
A list of matrices.
}
\description{
Extract the R-side variance components from a fitted model (asreml or lme)
and format them as matrices.
}
\examples{
\dontrun{
require('agridat')
dat <- hanks.sprinkler
dat <- transform(dat, subf=factor(subplot),
                 irrf=factor(irr))
dat <- dat[order(dat$block, dat$gen, dat$subplot),]
m1 <- asreml(yield ~ gen + dir + irrf + gen:dir + gen:irrf + dir:irrf,
             data=dat,
             random= ~ block + block:dir + block:irrf,
             rcov= ~ block:gen:corb(subf, k=3))
require("lucid")
lucid(rcor(m1)$subf[1:8,1:8],dig=2)
}
}
